refactor: reusable wsdl_downloader + generate SOAP bindings option
